Updating README to refine details
`git branch master` does not checkout the branch. We should also ensure
that we are deploying using an up to date code version; This is only
relevant if there are upstream changes to the deploy process. We need
to keep those in sync.

Capistrano runs the deploy script locally. It SSH-es into the app
server and pulls the changes from the remote repository (e.g. Github).