Adding auto-logging for click and visit activity
The goal of this addition is to automatically log the resulting URL of
all click, submit, and visit events.

This will in turn move the chatter of logging out of the scenario's
body, which should help improve the legibility. It will also

This is a step in addressing the general logging strategy that can
extend beyond the curate tests.