Adding the ./bin/run_tests script and docs
I want a more predictable entry point for our tests. As I've been
talking with the team, we may need a greater control of how we are
running the tests (e.g. run Curate against prod, but Sipity against
PPRD). By introducing a command, we begin to create that possibility.

It also pushes a procedure into code, which allows for documentation
at the point of usage (instead of referencing a README for all of the
nuances of running a test).