Begin logging examples to files
With this change, we begin logging the general chatter of the specs
along with the failures of those specs to the same log file.

At present, I am opting to write to both STDOUT and log files; This
way we can see how the specs look when they are run.

DLTP-747