Jenkins job support for webRenovation project
Jenkins job has an input parameter list for user to pick app name
to run tests against.

This variable is used by Capistrano as a server roles in
config/deploy/production.rb to determine which server the git repo
needs to be deployed to.

The same variable is used to set rspec command path in
run_test task