Adding capability to read swagger definition from private repositories
The viceroy API is a private repository and needs authorization
to read files like swagger definitions.
This implementation will use a personal access token generated
for one of our 'machine' users in github. This token will be read
from corpFS and passed in as a header value to the open() function.

This is a temporary implementation that will be consolidated with
Figaro in the next sprint.