Creating a new module to set Capybara.app_host
In this commit:
* I'm moving initialize_app_host to the InitializeExample module,
it used to sit in the ExampleLogging module before. I want to simplify
ExampleLogging to what we originally built it for.
* Adding if condition in initialize_app_host to evaluate value of
ENVIRONMENT and set Capybara.app_host to its value if valid_url? returns
true, otherwise it looks up the environment config file
* Calling the initialize_app_host method explicitly from config.before
hook in spec_helper for easier code readability