Using a wrapper around get_response
The Net::HTTP.get_response method does not support timeout. So in
case a website has a link that is unreachable the test aborts.
I want to continue using the get_response method because it returns a
Net::HTTPResponse object, whereas 'get' returns a string.
To meet both of these requirements I'm wrapping the get_response method
with the get method.