Returning a mocked up object for 'options' method
When the request builder encounters the 'options' method it was
returning to the scenario without any response object and causing the
scenario to fail in the response validator.
With this change I'm returning a mocked up response object of type
RestClient::Response and marking the scenario as pass in the
ResponseValidator class.