Basic Web Site Cases 8 & 9 These two scenarios have been included under a new function entitled User Navigation, and check the criteria of the tests against the site as much as they would match up. Some of the criteria for Basic 9 could not be met, as the "all Libraries" and "all about" buttons are no longer in the tabs associated with them.
I added two files as well to keep clutter down on the main web-ren file
for both of the tests, and have commited both of those as well.

The previous commit was for fixes for the database searching scenarios
that I had create earier, but that I forgot to check in the main webren
file.