Updates sipity functional test to use global LoginPage Instead of using the local login page, the sipity scenarios now use the global login page.
The submitted_etd.rb file was altered to account for a missing step
in testing the page by clicking the login button.

The cas_login global page now has terms_of_service_accepted as a
instance variable so that accounts which have accepted the terms and
the ones which have no are all tested.