Adding assertion on version number of usurper
Dev team has started injecting version number of the usurper application
as a meta element on every page of the site. This assertion helps us
make sure we are testing the right version of release.

* All usurper page classes are inherited from BasePage class and the
on_page? method of BasePage is asserted from every class. Adding this
version number assertion to BasePage class ensures that it's asserted
even when if we try to run a subset of scenarios
* Instead of asserting on the id and values in the meta tag, I'm built
the entire selector as a variable before asserting on it. This is the
only possible approach to overcome Capybara's limitations, it does not
support :content option in it's have_selector? method.