Accomodating assertion for appended '#' in URLs
To make the website more accessible, the dev team have added href='#'
to all the menu buttons.
Now for each of these pages, we're asserting for
Usurper::Pages::BasePage in which these navigation menu items are
being clicked and verified. This action appends a '#' at the
end of the URLs.

I added an either-or assertion for current_url so it still works
when objects of these pages are asserted for from other scenarios
that don't involve clicking the nav menu buttons.