Fixes condition when instance variable is nil
Not all specs are using the hook for excluding URIs from network
validation. In that case the instance variable
'@exclude_uri_from_network_traffic_validation' does not get created and
is 'nil'. The 'verify_network_traffic' method then fails with error:
```error
"class": "NoMethodError",
"message": "undefined method `include?' for nil:NilClass"
```

This fix adds an additional condition to check if the class variable
is not nil.