Adds AwsSsmHandler module for creating SSM clients
* This commit creates a new AwsSsmHandler module which is designed to
handle two scenarios:
1. Running on local, (RUNNING_ON_LOCAL_DEV=true)
In this scenario, the ssm client is created with an assumed IAM role.
This is necessary because secrets in AWS parameter store access
can only be read by a specific IAM role. A developer would not be
able to read them locally using their own AWS account

2. Running on AWS ECS, (RUNNING_ON_LOCAL_DEV=false)
In this scenario, the ssm client is created with an instance profile
credentials object that can be accessed by leveraging
Aws::InstanceProfileCredentials. This object contains credentials
of the EC2 instance in which the docker service is running in AWS, which
has permissions setup to read from the AWS parameter store

* This commit also exposes a `get_param_from_parameter_store` method
that can be used across other modules to read secrets from the parameter
store