Adds logic to handle aws creds/tokens from ENV
As part of dltp-1248, I want to ease the assume role setup process.
In most cases a developer would use a 3rd party or custom script
like [aws-vault](https://github.com/99designs/aws-vault) to setup
profiles. If that's the case, (which should be the preferred approach)
I can directly leverage the values of tokens and keys from ENV.