Removes link/button assertions for orcid page
Depending on how the testing accounts are setup across different
environments, the orcid redirect page can either be a sign in or registration
page. This keeps changing as we're testing different environments
and introduces unnecessary flux.

With this change I'm loosening up assertions to a point where it
verifies that the redirect was successful to either of the Orcid page.