feat: add puppeteer testing for e2e login checks
Due to the nature of the way SPAs are rendered in the browser, selenium doesn't work very well as a testing framework. This adds tests written in Puppeteer, which is a Node library and allows promises to be used as we wait for content to be rendered in the browser, making the front end tests more reliable. The tests are also written in a jest style format, making them as approachable as unit tests for some. Here, we add two basic tests for expected content - that the browser is going to the right place and that a login button that works is present - along with an end-to-end login check, if approrpiate ENV is leveraged. Instructions are in the README for both the Docker build/test and running the test suite without Docker.

Closes ESU-1142